+++ CHANGELOG.md
@@ -12,11 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `--debug` as an alias for `RUST_LOG=debug`.
 - Added `--trace` as an alias for `RUST_LOG=trace`.
 - Added notes about `--debug` and `--trace` when error occurs.
+- Added `left|right.folder.aliases` to define custom folder aliases.
 
 ### Changed
 
 - Replaced `anyhow` by [`color-eyre`](https://crates.io/crates/color-eyre) for better error management.
 - Replaced `log` by [`tracing`](https://crates.io/crates/tracing) for better log management.
+- Renamed `folder.filter` to `folder.filters` in order to match lib types.
+- Renamed `envelope.filter` to `envelope.filters` in order to match lib types.
 
 ## [0.1.0] - 2024-04-10